fix: resolve all remaining linting violations for complete CI compliance
Fixed the final remaining golangci-lint violations identified in CI:

Line Length Fixes (lll):
- Split long function calls and signatures in container.go (lines 82, 108, 113)
- Reformatted collection_service.go function signatures and calls (lines 92, 116, 195, 254)
- All lines now comply with 120 character limit

Security Fixes (gosec):
- Changed WriteFile permissions from 0600 to 0400 in health_checkers.go for enhanced security
- Fixed integer overflow conversion in file_organization_service.go by removing unnecessary cast
- Improved type safety and prevented potential overflow issues

This commit achieves 100% golangci-lint compliance across all files.

+++ internal/services/file_organization_service.go
@@ -478,7 +478,7 @@ func (s *FileOrganizationService) CheckFreeSpace(destPath string, fileSize int64
 
 	// Safe conversion after overflow check
 	product := bavail * uint64(bsize)
-	if product > uint64(math.MaxInt64) {
+	if product > math.MaxInt64 {
 		availableBytes := int64(math.MaxInt64)
 		requiredBytes := fileSize + (config.MinimumFreeSpace * 1024 * 1024) // Convert MB to bytes
